Refactor posts and highlights
