--- src/app/blogPosts.json
@@ -1,10 +1,0 @@
-{
-  "posts": [
-    {
-      "title": "My First Post",
-      "description": "This is my first blog post.",
-      "url": "/blog/2023/my-first-post",
-      "date": "2023-12-19"
-    }
-  ]
-}
