--- src/app/highlights.json
@@ -1,5 +1,0 @@
-{
-  "thoughts": [],
-  "articles": [],
-  "projects": []
-}
