+++ src/app/highlights.ts
@@ -1,0 +1,18 @@
+type Highlight = {
+  title: string;
+  description: string;
+};
+
+type Highlights = {
+  thoughts: Highlight[];
+  articles: Highlight[];
+  projects: Highlight[];
+};
+
+const highlights: Highlights = {
+  "thoughts": [],
+  "articles": [],
+  "projects": []
+};
+
+export default highlights;
